+++ packages/datadog-core/src/pkg.js
